Google Cloud Platform/requirements.txt
tensorflow==2.10.0
google-cloud-storage==1.16.1
Pillow==6.0.0
opencv-python


